fix: correct a subtitle in The Splitter
In #267 I fixed many errors in the subtitles for The Splitter. I *intended* to fix one where "n'est-ce pas" was written as "ne's pa", but apparently I mistakenly included an unmodified version of the `dlge.json`. This will correct that error, and the subtitle.

May Ne's pa rest in peace. I hope Ne's ma is doing well in these difficult times.